小黑板排产单权限设置以及修改金钱单位为2位数浮点型

+++ database/migrations/2020_05_22_025519_create_setting_table.php
@@ -23,7 +23,7 @@ class CreateSettingTable extends Migration
             $table->string('power');
             $table->integer('phases_number');
             $table->string('unit');
-            $table->integer('warehousing_price');
+            $table->float('warehousing_price',8,2);
         });
     }
